refactor: update import paths to use index.js for consistency
