+++ src/data/cache.ts
@@ -1,4 +1,4 @@
-import type { BaseDocument } from "../types/type.js";
+import type { BaseDocument } from "../types/index.js";
 
 export default class Cache<T = BaseDocument> {
 	#cache = new Map<string, T>();
