+++ src/data/metadata.ts
@@ -1,7 +1,7 @@
 import { readFile, stat, writeFile } from "node:fs/promises";
 import { basename, join } from "node:path";
 import AsyncMutex from "../utils/mutex.js";
-import type { CollectionMetadata } from "../types/type.js";
+import type { CollectionMetadata } from "../types/index.js";
 import { MetadataPersistenceError } from "../core/errors.js";
 export default class Metadata {
   #metadataPath: string;
